+++ src/interactivity/featureVizProperty.js
@@ -1,4 +1,4 @@
-import { blend, property, transition, notEquals } from '../renderer/viz/expressions';
+import { generateBlenderFunction, generateResetFunction } from './blendUtils';
 import { parseVizExpression } from '../renderer/viz/parser';
 
 /**
@@ -17,48 +17,11 @@ export default class FeatureVizProperty {
         this._properties = feature;
         this._viz = viz;
 
-        this.blendTo = _generateBlenderFunction(propertyName, feature[idProperty], customizedFeatures, viz, trackFeatureViz, idProperty);
-        this.reset = _generateResetFunction(propertyName, feature[idProperty], customizedFeatures, viz, idProperty);
+        this.blendTo = generateBlenderFunction(propertyName, feature[idProperty], customizedFeatures, viz, trackFeatureViz, idProperty, parseVizExpression);
+        this.reset = generateResetFunction(propertyName, feature[idProperty], customizedFeatures, viz, idProperty);
     }
 
     get value () {
         return this._viz[this._propertyName].eval(this._properties);
     }
-}
-
-function _generateResetFunction (propertyName, id, customizedFeatures, viz, idProperty) {
-    return function reset (duration = 500) {
-        if (customizedFeatures[id] && customizedFeatures[id][propertyName]) {
-            customizedFeatures[id][propertyName].replaceChild(
-                customizedFeatures[id][propertyName].mix,
-                // transition(0) is used to ensure that blend._predraw() "GC" collects it
-                blend(notEquals(property(idProperty), id), transition(0), transition(duration))
-            );
-            viz[propertyName].notify();
-            customizedFeatures[id][propertyName] = undefined;
-        }
-    };
-}
-
-function _generateBlenderFunction (propertyName, id, customizedFeatures, viz, trackFeatureViz, idProperty) {
-    return function generatedBlendTo (newExpression, duration = 500) {
-        if (typeof newExpression === 'string') {
-            newExpression = parseVizExpression(newExpression);
-        }
-        if (customizedFeatures[id] && customizedFeatures[id][propertyName]) {
-            customizedFeatures[id][propertyName].a.blendTo(newExpression, duration);
-            return;
-        }
-        const blendExpr = blend(
-            newExpression,
-            viz[propertyName],
-            blend(1, notEquals(property(idProperty), id), transition(duration))
-        );
-        trackFeatureViz(id, propertyName, blendExpr, customizedFeatures);
-        viz.replaceChild(
-            viz[propertyName],
-            blendExpr
-        );
-        viz[propertyName].notify();
-    };
 }